Huge Code review
There are too many things to list that I made in this commit -- my best advice is to
read through the changes I made instead of merging them. Here are some
things I did:

The first thing is rewrite some of your logic. I couldn't help it. Some
of your blocks were not the best.

The second thing is use the rust type system to its full advantage.
Meaning I made the type parameters to reflect the traits you implemented
and more.

The third thing is just general code style.